+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.2.0 (December 2, 2019)
+
+* Add support for `Bulk operations` feature (Create/Update/Delete and Query)
+* Add `Delete Object` action
+* Add `Lookup Objects` action
+* `Create object` action: add ability to utilize binary data attachment from previous step
+* `Upsert object` action: add ability to utilize binary data attachment from previous step
+* `Lookup Object (at most 1)` action: add ability to pass binary data (if found object has it) to the next component as a binary attachment
+* `Query` action: add ability to query deleted objects
+
 ## 1.1.2 (October 28, 2019)
 
 * Change Oauth values naming